==> Evaluate/notebooks/2_surveys_analysis.Rmd <==
---
title: "Crowd4SDG Surveys Analysis"
output: html_notebook
---

0. Load libraries and the dataframes for analysis

```{r}

library("ggplot2")
library("dplyr")
library("stringr")
library("plotly")
library("wesanderson")
library("harrypotter")

source("custom_functions.R")

load("../processed data/surveys.RData")
load("../processed data/registration.RData")

map = read.csv("../data/team_users_hashed.csv")
colnames(map) = c("Team", "ID", "Hash", "Mentors")

map = merge(map, reg[,c("ID","gender", "country_resid", "education")], by.x = "Hash", by.y = "ID", all.x = TRUE)

colnames(map) = c("hash","Team","ID","mentors","gender","residence","education")

set.seed(123)
```

1. Network of Interactions

```{r}

#User 34 is a special case - to confirm with Camille

temp = interactions[!interactions$to == 34, c(1,8,2:7)]

g_interactions = graph_from_data_frame(temp, vertices = map[,c(3,2,1,4:6)])

E(g_interactions)$weight = 1
g_simp_int = simplify(g_interactions)

g_sans_org = induced_subgraph(g_interactions, v = V(g_interactions)[!V(g_interactions)$Team == "Organizing Team"])

plt = visPlot(g_simp_int, directed = FALSE, nodesize = degree(g_simp_int), edgewidth = E(g_simp_int)$weight)
visSave(plt, "../figures/all_interactions.html")

pdf("../figures/all_interactions.pdf")
plotNetwork(g_simp_int, coms = V(g_simp_int)$Team)
dev.off()

```

1.a All networks (based on interaction type)

```{r}

pdf("../figures/interaction_networks.pdf")

for (i in unique(interactions$survey_field_id))
{
  
  E(g_interactions)$hidden = c(!E(g_interactions)$survey_field_id == i)
  #g_temp = subgraph.edges(g_interactions, eids = E(g_interactions)[E(g_interactions)$survey_field_id == i], delete.vertices = FALSE)
  
  plotNetwork(g_interactions, maine = i, coms = V(g_interactions)$Team)
}

#color = fish(n = length(unique(V(g_interactions)$Team)), option = "Koumansetta_rainfordi", end = 0.9)
color = pals::alphabet(length(unique(V(g_interactions)$Team)))
names(color) = unique(V(g_interactions)$Team)

#plot(NULL ,xaxt='n',yaxt='n',bty='n',ylab='',xlab='', xlim=0:1, ylim=0:1)
#legend(x=-1.5, y=-1.1, names(color), pch=21, col="#777777", pt.bg=color, pt.cex=2, cex=.8, bty="n", ncol=2)

plot(NULL ,xaxt='n',yaxt='n',bty='n',ylab='',xlab='', xlim=0:1, ylim=0:1)
legend("topleft", legend = names(color), pch=16, pt.cex=3, cex=1.5, bty='n',
    col = color)

dev.off()

```
Without ORG Team


```{r}

pdf("../figures/interaction_networks_sans_org.pdf")

for (i in unique(interactions$survey_field_id))
{
  
  vec = calculate_hidden(g_interactions, key = V(g_interactions)$name[V(g_interactions)$Team == "Organizing Team"])
  
  E(g_interactions)$hidden = c(!E(g_interactions)$survey_field_id == i) | vec
  
  plotNetwork(g_interactions, maine = i, coms = V(g_interactions)$Team, hide = "Organizing Team")
}

#color = fish(n = length(unique(V(g_interactions)$Team)), option = "Koumansetta_rainfordi", end = 0.9)
#color = pals::alphabet(length(unique(coms)))
#names(color) = unique(V(g_interactions)$Team)

plot(NULL ,xaxt='n',yaxt='n',bty='n',ylab='',xlab='', xlim=0:1, ylim=0:1)
legend("topleft", legend = names(color), pch=16, pt.cex=3, cex=1.5, bty='n',
    col = color)


dev.off()

```

2. Task Allocation Network

```{r}

nodes = data.frame(name = unique(tasks$user_id), type = TRUE)
nodes = rbind(nodes, data.frame(name = unique(tasks$tasks), type = FALSE))

g_task = graph_from_data_frame(tasks[,c(1,8,2:7)], vertices = nodes)
E(g_task)$weight = 1
g_task_simp = simplify(g_task)

plt = visPlot(g_task_simp, nodesize = degree(g_task_simp), communities = V(g_task_simp)$type+0, directed = FALSE, edgewidth = E(g_task_simp)$weight)
visSave(plt, file = "../figures/user_task_bipartite.html")

```

2.b Wordcloud

```{r, fig.height=10, fig.width=10}
library(wordcloud)

pdf("../figures/wordcloud_tasks.pdf", 10, 10)

temp = table(tasks$tasks)
task_freq = data.frame(task = names(temp), value = temp)

wc = wordcloud(words = task_freq$task, 
                  freq = task_freq$value.Freq, 
                  min.freq = 0,max.words=100, random.order=FALSE, rot.per=0, colors=brewer.pal(8, "Dark2")) 


dev.off()
```

Bar Plot of the same

```{r}

plt = ggplot(tasks, aes(y = forcats::fct_rev(forcats::fct_infreq(tasks)))) + geom_bar(stat = "count", fill = "steelblue") + theme_bw(base_size = 15) + xlab("Count") + ylab("") + theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
#    panel.border = element_blank(),
    panel.background = element_blank()) 

#ggplotly(plt)

ggsave(filename = "../figures/task_freq.png", plt, width = 15, height = 5)

```

Projected Networks

```{r}

temp = bipartite_projection(g_task_simp)
user_proj = temp[[2]]
task_proj = temp[[1]]

```

Visualize the projected networks

```{r}

pdf("../figures/projected_task_networks.pdf")

plotNetwork(user_proj, maine = "User Projection")
plotNetwork(task_proj, maine = "Task Projection")

dev.off()
```

***********************************************************************

Satisfaction Analysis

```{r}

surveys$survey_field_id = str_replace(surveys$survey_field_id, "Week 3: ", "")
surveys$survey_field_id = str_replace(surveys$survey_field_id, "Week 4: ", "")

surveys = surveys %>% rowwise() %>% mutate(week = str_split(survey_name, ":")[[1]][1])

#Merge teams

surveys = merge(surveys, map[,c("hash","Team")], by.x = "user_hash", by.y = "hash", all.x = TRUE)



```

```{r}

pdf("../figures/survey_satisfaction.pdf")

val = c("Very unhelpful" = -2, "Unhelpful" = -1, "Neither unhelpful nor helpful" = 0, "Helpful" = 1, "Very helpful" = 2)

surveys$content = factor(surveys$content, levels = c("Very unhelpful", "Unhelpful", "Neither unhelpful nor helpful", "Helpful", "Very helpful"))

surveys$content_val = val[surveys$content]

for (i in unique(surveys$survey_field_id))
{
  temp = surveys[surveys$survey_field_id == i,]
  
  plt = ggplot(temp, aes(y = content)) + geom_bar(stat = "count", width = 0.9) + scale_y_discrete(drop = FALSE) + theme_bw(base_size = 15) + xlab("Count") + ylab("") + ggtitle(str_split(i, "-")[[1]][1]) 
  
  print(plt)
}

```

```{r}

surveys_agg = surveys %>% group_by(survey_field_id) %>% summarise(mean = mean(content_val), se = se(content_val))

surveys_agg = surveys_agg %>% rowwise() %>% mutate(sid_short = strsplit(survey_field_id, "-")[[1]][1])


plt = ggplot(surveys_agg, aes(y = reorder(sid_short, mean), x = mean)) + geom_bar(stat = "identity", width = 0.9) + geom_errorbar(aes(xmin = mean-se, xmax = mean+se), width = 0) + theme_bw(base_size = 25) + xlab("") + ylab("")


ggsave(plt, filename = "../figures/survey_satisfaction.png", height = 10, width = 25)

```

Heatmap

```{r}

surveys_agg = surveys %>% group_by(survey_field_id, content) %>% summarise(count = n())
surveys_agg$content = factor(surveys_agg$content, levels = c("Very unhelpful", "Unhelpful", "Neither unhelpful nor helpful", "Helpful", "Very helpful"))
surveys_agg$survey_field_id = str_replace(surveys_agg$survey_field_id, '(.+)-.+', '\\1')


plt = ggplot(surveys_agg[!is.na(surveys_agg$content),]) + geom_tile(aes(x = content, y = survey_field_id, fill = count), lwd = 1.5, linetype = 1) + scale_fill_gradientn(colors = hp(n = 7, option = "Always")) + theme_bw(base_size = 35) + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) + ylab("") + xlab("") + scale_x_discrete(drop = FALSE) + geom_text(aes(x = content, y = survey_field_id, label = count), color = "white", size = 12)


ggsave(plt, filename = "../figures/survey_satisfaction_heatmap.png", height = 20, width = 30)

#ggplotly(plt)

```

************************************************************************

Task over Time Heatmap

```{r}

tasks = tasks %>% rowwise() %>% mutate(week = str_split(survey_name, ":")[[1]][1])

task_hm = tasks[,c("week","tasks")] %>% group_by(week, tasks) %>% summarise(count = n())

#Normalise - By week 

task_hm$week = factor(task_hm$week, levels = c("Weekly 1", "Weekly 2", "Weekly 3", "Weekly 4"))
task_hm$order = 0

peak = c()

for (i in unique(task_hm$tasks))
{
  temp = task_hm[task_hm$tasks == i,]
  
  task_hm$count[task_hm$tasks == i] = (temp$count/sum(temp$count))
  peak = temp$week[which.max(temp$count)]
  task_hm$order[task_hm$tasks == i] = peak
}

plt = ggplot(task_hm) + geom_tile(aes(x = week, y = reorder(tasks, -order), fill = count), lwd = 1.5, linetype = 1) + scale_fill_gradientn(colors = hp(n = 7, option = "Always")) + theme_bw() + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_blank()) + ylab("") + xlab("")

ggsave(plt, filename = "../figures/activity_heatmap.png", height = 7, width = 7)

```
No Normalization

```{r}

tasks = tasks %>% rowwise() %>% mutate(week = str_split(survey_name, ":")[[1]][1])

task_hm = tasks[,c("week","tasks")] %>% group_by(week, tasks) %>% summarise(count = n())

#Normalise - By week 

task_hm$week = factor(task_hm$week, levels = c("Weekly 1", "Weekly 2", "Weekly 3", "Weekly 4"))
task_hm$order = 0

peak = c()

for (i in unique(task_hm$tasks))
{
  temp = task_hm[task_hm$tasks == i,]
  
#  task_hm$count[task_hm$tasks == i] = (temp$count/sum(temp$count))
  peak = temp$week[which.max(temp$count)]
  task_hm$order[task_hm$tasks == i] = peak
}

plt = ggplot(task_hm) + geom_tile(aes(x = week, y = reorder(tasks, -order), fill = count), lwd = 1.5, linetype = 1) + scale_fill_gradientn(colors = hp(n = 7, option = "Always")) + theme_bw() + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_blank()) + ylab("") + xlab("") + geom_text(aes(x = week, y = reorder(tasks, -order), label = count), color = "white")

ggsave(plt, filename = "../figures/activity_heatmap_nonor.png", height = 7, width = 7)

```

No Normalization and No ordering by time

```{r}

tasks = tasks %>% rowwise() %>% mutate(week = str_split(survey_name, ":")[[1]][1])

task_hm = tasks[,c("week","tasks")] %>% group_by(week, tasks) %>% summarise(count = n())

#Normalise - By week 

task_hm$week = factor(task_hm$week, levels = c("Weekly 1", "Weekly 2", "Weekly 3", "Weekly 4"))

plt = ggplot(task_hm) + geom_tile(aes(x = week, y = reorder(tasks, count), fill = count), lwd = 1.5, linetype = 1) + scale_fill_gradientn(colors = hp(n = 7, option = "Always")) + theme_bw(base_size = 15) + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_blank()) + ylab("") + xlab("") + geom_text(aes(x = week, y = reorder(tasks, count), label = count), color = "white") + ggtitle("Evaluate")

ggsave(plt, filename = "../figures/activity_heatmap_nonor_noroder.png", height = 7, width = 10)

```

Heatmaps by Team

```{r}

#New Team Names

reg_map = c("A2: Women & Technology Against Climate Change" = "T6: Women & Technology Against Climate Change", "B2: TEAM FOILED" = "T3: TEAM FOILED", "C1: Andapé Institute" = "T13: Andapé Institute", "C3: WOMER" = "T5: WOMER", "A5: Donate Water Project" = "T9: DonateWater", "B5: Rights of Climate" = "T11: Rights of Climate", "B3: Eco Winners" = "T14: Eco Winners", "B4: Women 4 Sustainable World" = "T12: Women 4 Sustainable World", "A1: Up Get App/CitiCERN" = "T7: UpGet app - CitiCERN Project", "B1: Water Warriors" = "T10: Water Warriors", "C2: PAM" = "T4: PAM", "C4: Climate Gender Justice" = "T8: Climate Gender Justice", "A3: Rhythm of Bamboos" = "T1: SDesiGn (Old name: Rhythm of Bamboos)", "C5: Ashifa Nazrin" = "C5: Ashifa Nazrin", "A4: Flood Rangers" = "T2: Flood Rangers")

reg_map = data.frame(old_name = names(reg_map), new_name = reg_map)
map = merge(map, reg_map, by.x = "Team", by.y = "old_name", all.x = TRUE)

tasks = merge(tasks, map[,c("hash", "new_name")], by.x = "user_hash", by.y = "hash")

pdf("../figures/tasks_per_team.pdf")

for (i in unique(tasks$tasks))
{
  temp = tasks[tasks$tasks == i,]
  
  task_hm = temp[,c("week","new_name")] %>% group_by(week, new_name) %>% summarise(count = n())

#Normalise - By week 

  task_hm$week = factor(task_hm$week, levels = c("Weekly 1", "Weekly 2", "Weekly 3", "Weekly 4"))
  task_hm$order = 0
  
  peak = c()
  
  for (j in unique(task_hm$new_name))
  {
    temp = task_hm[task_hm$new_name == j,]
    
    task_hm$count[task_hm$new_name == j] = (temp$count/sum(temp$count))
    peak = temp$week[which.max(temp$count)]
    task_hm$order[task_hm$new_name == j] = peak
  }
  
  plt = ggplot(task_hm) + geom_tile(aes(x = week, y = reorder(new_name, -order), fill = count), lwd = 1.5, linetype = 1) + scale_fill_gradientn(colors = hp(n = 7, option = "Always")) + theme_bw() + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank()) + ylab("") + xlab("") + ggtitle(i) + scale_x_discrete(drop=FALSE) + scale_y_discrete(drop=FALSE)
  
  print(plt)
  
}

dev.off()


```

Team Attendance in Surveys

```{r}

final = read.csv("../data/Crowd4SDG_gear2_final.csv")
final$week = "Final"
temp = final[ ,c("Username", "week", "What.is.your.team.")]
colnames(temp) = c("user_hash", "week", "Team")

temp$Team[temp$Team == "A2: Women & Technology Against Climate Change"] = "T6: Women & Technology Against Climate Change"

att_surveys = surveys[,c("user_hash", "week")]
att_surveys = merge(att_surveys, map[,c("hash", "new_name")], by.x = "user_hash", by.y = "hash", all.x = TRUE)
colnames(att_surveys) = c("user_hash", "week", "Team")

att_surveys = rbind(att_surveys, temp)


att_surveys = att_surveys %>% group_by(Team, week) %>% summarise(count = if (n() > 0) 1 else 0)
att_surveys$week = factor(att_surveys$week, levels = c("Weekly 1", "Weekly 2", "Weekly 3", "Weekly 4", "Final"))

plt = ggplot(att_surveys, aes(x = week, y = Team, fill = count)) + geom_tile(lwd = 1.5, linetype = 1) + theme_bw(base_size = 20) + scale_fill_gradientn(colors = hp(n = 1, option = "Always"), na.value = "white") + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank()) + ylab("") + xlab("")

ggsave(plt, filename = "../figures/team_survey_attendance_weekly.png", width = 12, height = 7)

#ggplotly(plt)

```
